Split interruption and parJoin
